requirements.txt
numpy==1.26.3
pandas==1.5.2
matplotlib==3.8.2
scikit-learn==1.3.2
Flask==1.1.2
Flask-RESTful==0.3.8
MarkupSafe==1.1.1
itsdangerous==1.1.0
tensorflow==2.15.0
opencv-python==4.9.0.80
Werkzeug==1.0.1
Jinja2==2.11.2
click==7.1.2
reparse==3.0
Pillow